OK-634: Move Europass stuff into its own package
